lynx device branch

+++ manage.sh
@@ -35,6 +35,7 @@ readonly aosp_forks=(
     device_google_gs101-sepolicy
     device_google_gs201
     device_google_gs201-sepolicy
+    device_google_lynx
     device_google_pantah
     device_google_raviole
     device_google_redbull
@@ -162,6 +163,7 @@ readonly independent=(
     carriersettings-extractor
     device_google_bluejay-kernel
     device_google_coral-kernel
+    device_google_lynx-kernel
     device_google_pantah-kernel
     device_google_raviole-kernel
     device_google_redbull-kernel
@@ -173,6 +175,7 @@ readonly independent=(
     kernel_manifest-5.15
     kernel_manifest-bluejay
     kernel_manifest-coral
+    kernel_manifest-lynx
     kernel_manifest-pantah
     kernel_manifest-raviole
     kernel_manifest-redbull
@@ -215,8 +218,16 @@ for repo in "${aosp_forks[@]}"; do
         fi
     elif [[ $action == update ]]; then
         git fetch upstream --tags
-        git rebase --onto $aosp_tag $aosp_tag_old
-        git push -f
+
+        if [[ $repo != @(platform_build|platform_manifest|device_google_lynx) ]]; then
+            git checkout $aosp_tag
+            git cherry-pick $aosp_base_tag..$base_branch
+            git checkout -B 13-lynx
+            git push -fu origin 13-lynx
+        else
+            git rebase --onto $aosp_tag $aosp_tag_old
+            git push -f
+        fi
     elif [[ $action == push ]]; then
         git push
     elif [[ $action == fetch ]]; then
@@ -227,10 +238,12 @@ for repo in "${aosp_forks[@]}"; do
 done
 
 for repo in ${kernels[@]}; do
+    [[ $action != @(release|delete) ]] && continue
+
     echo -e "\n>>> $(tput setaf 3)Handling $repo$(tput sgr0)"
 
     cd $repo
-    git checkout $branch
+    git checkout $base_branch
 
     if [[ $action == delete ]]; then
         git tag -d $delete_tag || true
@@ -261,14 +274,20 @@ for repo in ${independent[@]}; do
         git tag -d $delete_tag || true
         git push origin --delete $delete_tag || true
     elif [[ $action == release ]]; then
-        if [[ $repo == @(kernel_manifest-5.10|kernel_manifest-5.15|kernel_manifest-bluejay|kernel_manifest-coral|kernel_manifest-pantah|kernel_manifest-redbull|kernel_manifest-raviole) ]]; then
+        if [[ $repo == @(kernel_manifest-5.10|kernel_manifest-5.15|kernel_manifest-bluejay|kernel_manifest-coral|kernel_manifest-lynx|kernel_manifest-pantah|kernel_manifest-redbull|kernel_manifest-raviole) ]]; then
             git checkout -B tmp
-            sed -i s%refs/heads/$branch%refs/tags/$aosp_version.$build_number% default.xml
+            sed -i s%refs/heads/$base_branch%refs/tags/$aosp_version.$build_number% default.xml
             git commit default.xml -m $aosp_version.$build_number
             git push -fu origin tmp
         else
             git tag -s $aosp_version.$build_number -m $aosp_version.$build_number
             git push origin $aosp_version.$build_number
+        fi
+    elif [[ $action == update ]]; then
+        if [[ $repo != @(device_google_lynx-kernel|kernel_manifest-lynx|script) ]]; then
+            git checkout $base_branch
+            git checkout -B $branch
+            git push -fu origin $branch
         fi
     elif [[ $action == push ]]; then
         git push
